+++ pkg/server/handler.go
@@ -1,0 +1,5 @@
+package server
+
+func (server *Server) RegisterHandler() {
+	//router := server.app.Group("v1")
+}