+++ pkg/server/middleware.go
@@ -1,0 +1,7 @@
+package server
+
+import "github.com/gofiber/fiber/v2"
+
+func (server *Server) Use(handler fiber.Handler) {
+	server.app.Use(handler)
+}