+++ pkg/server/server.go
@@ -1,0 +1,15 @@
+package server
+
+import (
+	"github.com/gofiber/fiber/v2"
+)
+
+type Server struct {
+	app *fiber.App
+}
+
+func New(config ...fiber.Config) *Server {
+	return &Server{
+		app: fiber.New(config...),
+	}
+}